[Issue-1012] [Keyring] Update routing
