+++ packages/extension-koni-ui/src/Popup/Account/ImportQrCode.tsx
@@ -5,9 +5,9 @@ import ChainLogoMap from '@subwallet/extension-koni-ui/assets/logo';
 import { Layout } from '@subwallet/extension-koni-ui/components';
 import DualLogo from '@subwallet/extension-koni-ui/components/Logo/DualLogo';
 import QrScannerErrorNotice from '@subwallet/extension-koni-ui/components/Qr/Scanner/ErrorNotice';
+import useCompleteCreateAccount from '@subwallet/extension-koni-ui/hooks/account/useCompleteCreateAccount';
 import useGetDefaultAccountName from '@subwallet/extension-koni-ui/hooks/account/useGetDefaultAccountName';
 import useAutoNavigateToCreatePassword from '@subwallet/extension-koni-ui/hooks/router/autoNavigateToCreatePassword';
-import useDefaultNavigate from '@subwallet/extension-koni-ui/hooks/router/useDefaultNavigate';
 import { checkPublicAndPrivateKey, createAccountWithSecret } from '@subwallet/extension-koni-ui/messaging';
 import { ThemeProps } from '@subwallet/extension-koni-ui/types';
 import { QrAccount } from '@subwallet/extension-koni-ui/types/scanner';
@@ -53,7 +53,7 @@ const Component: React.FC<Props> = (props: Props) => {
 
   const { className } = props;
   const { t } = useTranslation();
-  const { goHome } = useDefaultNavigate();
+  const onComplete = useCompleteCreateAccount();
   const accountName = useGetDefaultAccountName();
 
   const { activeModal, inactiveModal } = useContext(ModalContext);
@@ -94,7 +94,7 @@ const Component: React.FC<Props> = (props: Props) => {
               if (success) {
                 inactiveModal(modalId);
                 setValidateState({});
-                goHome();
+                onComplete();
               } else {
                 setValidateState({
                   message: errors[0].message,
@@ -121,7 +121,7 @@ const Component: React.FC<Props> = (props: Props) => {
     } else {
       setLoading(false);
     }
-  }, [account, accountName, goHome, inactiveModal]);
+  }, [account, accountName, onComplete, inactiveModal]);
 
   const openCamera = useCallback(() => {
     activeModal(modalId);
